@@ -1,10 +1,12 @@
-import React from "react";
-import "./Container.css";
+import React from 'react';
+import './Container.css';
 
-// import { Container } from './styles';
+interface Props {
+  children: JSX.Element[] | JSX.Element;
+}
 
-const Container: React.FC = (props) => {
-  return <div className="AppContainer">{props.children}</div>;
+const Container: React.FC<Props> = (props) => {
+  return <div className='AppContainer'>{props.children}</div>;
 };
 
 export default Container;
